Structures: Move OCDevAddr to its own file
